Refactor app.py into modular structure with correct libraries
- Split app.py into main.py, utils/tools.py, utils/prompts.py, scripts/init_db.py
- Updated libraries to strands-agents and strands-agents-tools
- Moved init_db to scripts/init_db.py as a prerequisite
- Implemented CLI for list-tables, explain-query, create-bank-table, fill-bank-table
- Updated pyproject.toml with requires-python '>=3.10' and correct dependencies
- Added tests/test_tools.py for tool validation
- Updated README.md with setup instructions and architecture diagram
