+++ 02-samples/07-Multi-Agent-datawarehouse-Query-Optimiser/pyproject.toml
@@ -1,0 +1,11 @@
+[project]
+name = "query-optimizer-swarm"
+version = "0.1.0"
+description = "Multi-Agent Data Warehouse Query Optimizer using SQLite and AWS Bedrock"
+requires-python = ">=3.10"
+dependencies = [
+    "strands-agents",
+    "strands-agents-tools",
+    "boto3",
+    "python-dotenv"
+]
