+++ 02-samples/07-Multi-Agent-datawarehouse-Query-Optimiser/scripts/__init__.py
@@ -1,0 +1,1 @@
+"""Scripts for prerequisites."""
