+++ 02-samples/07-Multi-Agent-datawarehouse-Query-Optimiser/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the Multi-Agent Data Warehouse Query Optimizer."""
